feat: raw encoding/decoding
Encodes using `O = (I+42) % 256` formula. If O is a special characters,
it is prefixed with the escape character (`=`), while increased by 64,
modulo 256.

Supports encoding single byte, and binary.

The decoding process reverses encoding process, and is simply tested by
symmetric property check. There are properties for the encoding process,
so the symmetric property should be correct.